+++ src/app/modules/marketplace/components/module-train/storage-conf-form/storage-conf-form.component.ts
@@ -1,17 +1,30 @@
 import { MediaMatcher } from '@angular/cdk/layout';
 import { ChangeDetectorRef, Component, Input, OnInit } from '@angular/core';
-import { FormBuilder, FormGroup, FormGroupDirective } from '@angular/forms';
+import {
+    FormBuilder,
+    FormControl,
+    FormGroup,
+    FormGroupDirective,
+    Validators,
+} from '@angular/forms';
 import {
     ModuleStorageConfiguration,
-    confObjectRange,
+    confObject,
+    confObjectStringBoolean,
 } from '@app/shared/interfaces/module.interface';
+import { ZenodoSimpleDataset } from '@app/shared/interfaces/dataset.interface';
 
-const mockedConfObject: confObjectRange = {
-    range: [],
+const mockedConfObject: confObject = {
     name: '',
     value: '',
     description: '',
 };
+const mockedConfObjectStringBoolean: confObjectStringBoolean = {
+    name: '',
+    value: { stringValue: '', booleanValue: false },
+    description: '',
+};
+
 @Component({
     selector: 'app-storage-conf-form',
     templateUrl: './storage-conf-form.component.html',
@@ -37,6 +50,11 @@ export class StorageConfFormComponent implements OnInit {
         rcloneVendorSelect: [''],
         rcloneUserInput: [''],
         rclonePasswordInput: [''],
+        zenodoCommunitySelect: new FormControl({ value: '', disabled: true }),
+        zenodoDatasetSelect: new FormControl({ value: '', disabled: true }),
+        zenodoVersionSelect: new FormControl({ value: '', disabled: true }),
+        doiInput: [''],
+        datasetsList: [[{ doi: '', force_pull: false }]],
     });
 
     protected _defaultFormValues: ModuleStorageConfiguration = {
@@ -45,12 +63,8 @@ export class StorageConfFormComponent implements OnInit {
         rclone_vendor: mockedConfObject,
         rclone_user: mockedConfObject,
         rclone_password: mockedConfObject,
+        datasets: mockedConfObjectStringBoolean,
     };
-
-    protected _showHelp = false;
-
-    mobileQuery: MediaQueryList;
-    private _mobileQueryListener: () => void;
 
     @Input() set showHelp(showHelp: boolean) {
         this._showHelp = showHelp;
@@ -64,12 +78,15 @@ export class StorageConfFormComponent implements OnInit {
             this.storageConfFormGroup
                 .get('rcloneConfInput')
                 ?.setValue(defaultFormValues.rclone_conf.value as string);
+
             this.storageConfFormGroup
                 .get('storageUrlInput')
+
                 ?.setValue(defaultFormValues.rclone_url.value as string);
             this.storageConfFormGroup
                 .get('rcloneUserInput')
                 ?.setValue(defaultFormValues.rclone_user.value as string);
+
             this.storageConfFormGroup
                 .get('rcloneVendorSelect')
                 ?.setValue(defaultFormValues.rclone_vendor.value as string);
@@ -81,11 +98,17 @@ export class StorageConfFormComponent implements OnInit {
                     });
                 }
             );
+            this.storageConfFormGroup.get('datasetsList')?.setValue([]);
         }
     }
 
+    protected _showHelp = false;
+    private _mobileQueryListener: () => void;
+    mobileQuery: MediaQueryList;
+
     hidePassword = true;
     rcloneVendorOptions: { value: string; viewValue: string }[] = [];
+    datasets: { doi: string; force_pull: boolean }[] = [];
 
     ngOnInit(): void {
         this.parentForm = this.ctrlContainer.form;
@@ -93,5 +116,44 @@ export class StorageConfFormComponent implements OnInit {
             'storageConfForm',
             this.storageConfFormGroup
         );
+    }
+
+    addDataset(dataset: ZenodoSimpleDataset): void {
+        const rcloneUser = this.storageConfFormGroup.get('rcloneUserInput');
+        const rclonePassword = this.storageConfFormGroup.get(
+            'rclonePasswordInput'
+        );
+        rcloneUser?.markAllAsTouched();
+        rclonePassword?.markAllAsTouched();
+        rcloneUser?.setValidators([Validators.required]);
+        rclonePassword?.setValidators([Validators.required]);
+        rcloneUser?.updateValueAndValidity();
+        rclonePassword?.updateValueAndValidity();
+
+        this.datasets.push({ doi: String(dataset.doi), force_pull: false });
+        this.storageConfFormGroup.get('datasetsList')?.setValue(this.datasets);
+    }
+
+    deleteDataset(dataset: ZenodoSimpleDataset): void {
+        this.datasets = this.datasets.filter((d) => d.doi !== dataset.doi);
+        this.storageConfFormGroup.get('datasetsList')?.setValue(this.datasets);
+
+        if (this.datasets.length == 0) {
+            const rcloneUser = this.storageConfFormGroup.get('rcloneUserInput');
+            const rclonePassword = this.storageConfFormGroup.get(
+                'rclonePasswordInput'
+            );
+            rcloneUser?.setValidators(null);
+            rclonePassword?.setValidators(null);
+            rcloneUser?.updateValueAndValidity();
+            rclonePassword?.updateValueAndValidity();
+        }
+    }
+
+    updateDataset(dataset: ZenodoSimpleDataset): void {
+        const d = this.datasets.find((d) => d.doi === dataset.doi);
+        if (d) {
+            d.force_pull = dataset.force_pull;
+        }
     }
 }